Liu4

+++ meteva/method/space/acc/acc.py
@@ -10,26 +10,29 @@ import urllib.request
 import traceback
 import math
 
-def acc(grd_ob,grd_fo):
+def acc(grd_ob,grd_fo,climate_path = None):
     grid0 = meteva.base.get_grid_of_data(grd_ob)
-    z500_cli_path = meteva.base.z500_cli
-    if not os.path.exists(z500_cli_path):
-        url = "https://github.com/nmcdev/meteva/tree/master/meteva/resources/z500_cli.nc"
-        try:
-            print("开始从github下载500hPa高度场气候平均数据，请稍等")
-            urllib.request.urlretrieve(url, filename=z500_cli_path)
-            z500_cli = meteva.base.read_griddata_from_nc(z500_cli_path, grid=grid0)
-            if z500_cli is None:
-                raise ValueError("z500_cli 不能是None")
-            print("下载500hPa高度场气候平均数据下载成功")
-        except Exception as e:
-            print("从github下载地形高度数据失败，请重试，或者手动从\n" + url + "\n下载文件并保存至\n" + z500_cli_path)
-            print(e)
-            if os.path.exists(z500_cli_path):
-                os.remove(z500_cli_path)
-            return
+    if climate_path is None:
+        z500_cli_path = meteva.base.z500_cli
+        if not os.path.exists(z500_cli_path):
+            url = "https://github.com/nmcdev/meteva/tree/master/meteva/resources/z500_cli.nc"
+            try:
+                print("开始从github下载500hPa高度场气候平均数据，请稍等")
+                urllib.request.urlretrieve(url, filename=z500_cli_path)
+                z500_cli = meteva.base.read_griddata_from_nc(z500_cli_path, grid=grid0)
+                if z500_cli is None:
+                    raise ValueError("z500_cli 不能是None")
+                print("下载500hPa高度场气候平均数据下载成功")
+            except Exception as e:
+                print("从github下载地形高度数据失败，请重试，或者手动从\n" + url + "\n下载文件并保存至\n" + z500_cli_path)
+                print(e)
+                if os.path.exists(z500_cli_path):
+                    os.remove(z500_cli_path)
+                return
+        else:
+            grd_cli_all =  meteva.base.read_griddata_from_nc(z500_cli_path, grid=grid0)
     else:
-        z500_cli =  meteva.base.read_griddata_from_nc(z500_cli_path, grid=grid0)
+        grd_cli_all = meteva.base.read_griddata_from_nc(climate_path, grid=grid0)
 
     time_ob = meteva.base.all_type_time_to_datetime(grid0.gtime[0])
     if time_ob.hour ==0 or time_ob.hour == 12:
@@ -38,7 +41,7 @@ def acc(grd_ob,grd_fo):
         time_ob_ut = time_ob - datetime.timedelta(hours=8)
 
     time_cli = datetime.datetime(2020, time_ob_ut.month, time_ob_ut.day, time_ob_ut.hour)
-    grd_cli =  meteva.base.in_time_list(z500_cli, [time_cli])
+    grd_cli =  meteva.base.in_time_list(grd_cli_all, [time_cli])
     grd_obs = copy.deepcopy(grd_ob)
     grd_obs.values -= grd_cli.values
     grd_fst = copy.deepcopy(grd_fo)
@@ -58,16 +61,28 @@ def acc(grd_ob,grd_fo):
 
 
 def acc_middle_z500(para_acc):
-    z500_cli_path = meteva.base.z500_cli
-    if not os.path.exists(z500_cli_path):
-        url = "https://github.com/nmcdev/meteva/raw/master/meteva/resources/z500_cli.nc"
-        try:
-            print("开始从github下载地形高度数据，请稍等")
-            urllib.request.urlretrieve(url, filename=z500_cli_path)
-        except Exception as e:
-            print("从github下载地形高度数据失败，请重试，或者手动从\n" + url + "\n下载文件并保存至\n" + z500_cli_path)
-            print(e)
-            return
+
+    para_acc["climate_path"] = meteva.base.z500_cli
+    acc_middle(para_acc)
+
+def acc_middle(para_acc):
+
+    if "climate_path" not in para_acc.keys():
+        para_acc["climate_path"] = None
+
+    if para_acc["climate_path"] is None:
+        climate_path = meteva.base.z500_cli
+        if not os.path.exists(climate_path):
+            url = "https://github.com/nmcdev/meteva/raw/master/meteva/resources/z500_cli.nc"
+            try:
+                print("开始从github下载地形高度数据，请稍等")
+                urllib.request.urlretrieve(url, filename=climate_path)
+            except Exception as e:
+                print("从github下载地形高度数据失败，请重试，或者手动从\n" + url + "\n下载文件并保存至\n" + climate_path)
+                print(e)
+                return
+    else:
+        climate_path = para_acc["climate_path"]
 
     middle_result_path = para_acc["middle_result_path"]
     model_name_list =list(para_acc["fo_data"].keys()) # ["ECMWF","PANGU","FENGWU"]  #数据的名称
@@ -75,17 +90,24 @@ def acc_middle_z500(para_acc):
         marker = meteva.perspact.get_grid_marker(para_acc["grid"],step=para_acc["step"])
     else:
         marker = None
-    z500_cli =  meteva.base.read_griddata_from_nc(z500_cli_path,grid = para_acc["grid"])
+    grd_cli_all =  meteva.base.read_griddata_from_nc(climate_path,grid = para_acc["grid"])
 
     tmmsss_list = []
     #创建一个空的DataFrame，用来记录哪些中间量已经收集
     df1 = pd.DataFrame(data=None,columns=['level', 'time', 'dtime', 'member'])
     if os.path.exists(middle_result_path):
-        tmmsss0 = pd.read_hdf(middle_result_path)
+        _,file_type = os.path.splitext(middle_result_path)
+        if file_type=="parquet":
+            tmmsss0 = pd.read_parquet(middle_result_path)
+        else:
+            tmmsss0 = pd.read_hdf(middle_result_path)
+
         tmmsss_list.append(tmmsss0)
         df1 = tmmsss0[["level","time","dtime","member"]]   #如果有文件就替换掉df1
         df1.drop_duplicates(keep="first", inplace=True)   # 用来记录哪些时次已经收集了中间量
 
+    if "time_type" not in para_acc.keys():
+        para_acc["time_type"] = "BT"
 
     for model in model_name_list:
         df2 = df1.loc[df1['member'].isin([model])]
@@ -98,34 +120,43 @@ def acc_middle_z500(para_acc):
             if len(df3.index) ==len(dtime_list):
                 print(meteva.base.get_path(model + "YYYY年MM月DD日HH时起报的检验中间量已存在",time1))
             else:
-                time_ut = time1 - datetime.timedelta(hours=8)
+                if para_acc["time_type"] =="BT":
+                    time_ut = time1 - datetime.timedelta(hours=8)
+                else:
+                    time_ut = time1
+
                 for dh in dtime_list:
                     df4 =  meteva.base.in_dtime_list(df3,dtime_list=dh)
                     if len(df4.index) ==0:
                         save_k += 1  #
                         time_ob = time1 + datetime.timedelta(hours=dh)
-                        time_ob_ut = time_ob - datetime.timedelta(hours=8)
-                        if para_acc["ob_data"]["time_type"] =="BT":
-                            time_path = time_ob
-                        else:
+                        time_ob_ut = time_ut + datetime.timedelta(hours=dh)
+
+                        if para_acc["ob_data"]["time_type"] =="UT":
                             time_path = time_ob_ut
+                        else:
+                            time_path = time_ob_ut + datetime.timedelta(hours=8)
+
                         time_cli = datetime.datetime(2020,time_ob_ut.month,time_ob_ut.day,time_ob_ut.hour)
-                        grd_cli =  meteva.base.in_time_list(z500_cli,[time_cli])
+                        grd_cli =  meteva.base.in_time_list(grd_cli_all,[time_cli])
 
                         path0 =  meteva.base.get_path(para_acc["ob_data"]["dir_ob"], time_path)
-                        grd_obs = para_acc["ob_data"]["read_method"](path0, grid=para_acc["grid"], time=time_ob, dtime=0, data_name="OBS",show = True,**para_acc["ob_data"]["read_para"])
+                        grd_obs = para_acc["ob_data"]["read_method"](path0, grid=para_acc["grid"], time=time_path, dtime=0, data_name="OBS",show = True,**para_acc["ob_data"]["read_para"])
                         if grd_obs is None: continue
+                        meteva.base.set_griddata_coords(grd_obs,gtime=[time_ob])
 
                         grd_obs.values *= para_acc["ob_data"]["multiple"]
                         grd_obs.values -= grd_cli.values
 
-                        if para_fo1["time_type"]=="BT":
-                            path1 =  meteva.base.get_path(para_fo1["dir_fo"],time1,dh)
+                        if para_fo1["time_type"]=="UT":
+                            time_path = time_ut
                         else:
-                            path1 =  meteva.base.get_path(para_fo1["dir_fo"], time_ut, dh)
+                            time_path = time_ut + datetime.timedelta(hours=8)
 
-                        grd_fo = para_fo1["read_method"](path1, grid=para_acc["grid"], time=time1, dtime=dh, data_name=model,show = True,**para_fo1["read_para"])
+                        path1 = meteva.base.get_path(para_fo1["dir_fo"], time_path, dh)
+                        grd_fo = para_fo1["read_method"](path1, grid=para_acc["grid"], time=time_path, dtime=dh, data_name=model,show = True,**para_fo1["read_para"])
                         if grd_fo is not None:
+                            meteva.base.set_griddata_coords(grd_fo, gtime=[time1])
                             grd_fo.values *=para_fo1["multiple"]
                             grd_fo.values -= grd_cli.values
                             df_mid =  meteva.perspact.middle_df_grd(grd_obs, grd_fo, meteva.method.tmmsss, marker=marker)
@@ -145,7 +176,13 @@ def acc_middle_z500(para_acc):
     # 先删除文件在重新输出文件，避免文件大小膨胀
     if os.path.exists(middle_result_path):
         os.remove(middle_result_path)
-    tmmss_all.to_hdf(middle_result_path, "df")  # 将结果输出到文件
+
+    file_type = middle_result_path.split(".")[1]
+    if file_type =="parquet":
+        tmmss_all.to_parquet(middle_result_path)  # 将结果输出到文件
+    else:
+        tmmss_all.to_hdf(middle_result_path, "df")  # 将结果输出到文件
+
     print("中间量统计程序运行完毕")
     return
 
@@ -156,8 +193,10 @@ if __name__=="__main__":
     para_acc = {
         "begin_time": datetime.datetime(2022,5,1,8),  #起始时间（预报）
         "end_time":  datetime.datetime(2022,5,5,8), #结束时间（预报）
+        "time_type":"UT",
         "time_step":12,     #起报时间间隔
-        "middle_result_path":  r"H:\test_data\output\method\space\acc\tmmsss_z500.h5",
+        "middle_result_path":  r"H:\test_data\output\method\space\acc\tmmsss_z500.parquet",
+        "climate_path":None,
         "grid" :  meteva.base.grid([70,140,0.25],[0,60,0.25]),  # 检验区域
         "step":2,  #masker间隔（单位：°）
         "ob_data": {
@@ -188,7 +227,7 @@ if __name__=="__main__":
     }
 
 
-    #acc_middle_z500(para_acc)
+    acc_middle(para_acc)
     df_mid = pd.read_hdf(para_acc["middle_result_path"])
     print(df_mid["member"])
     score,gdict = meteva.perspact.score_df(df_mid,meteva.method.corr,g = ["member","dtime"],plot = "line",save_path = r"H:\test_data\output\method\space\acc\acc_dtime.png",
